requirements.txt
# Google Gmail API
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.120.0

flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
python-dotenv>=1.0.0
langchain>=0.2.0
langchain-groq>=0.1.0
langgraph>=0.1.0
langchain-community>=0.2.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
openpyxl>=3.1.0
scikit-learn>=1.3.0
numpy>=1.24.0
sqlalchemy>=2.0.0
python-dateutil>=2.8.0
imaplib2>=3.6
secure-smtplib>=0.1.1
eventlet>=0.35.0
celery[redis]>=5.4.0
redis>=5.0.0
